Revert "Ano vi tfor mnist1"
