+++ .gitignore
@@ -1,12 +1,2 @@
-data/Lung_dataset_original/ 
-
-# Python バイトコード
-__pycache__/
-*.py[cod]
-
-# mypy・pytest 等のキャッシュも不要なら
-.mypy_cache/
-.pytest_cache/
-
-# ---------------- 環境ファイル ----------------
-# docker/.env
+/data/
+.env
