+++ AnoViT/test_bunpu_2D_theres.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 import os
 import pickle
 import numpy as np
