+++ docker/.env
@@ -1,15 +1,6 @@
 # データファイルのパスを指定
-#######################################################################
-#  Git 追跡を一時的に外す／戻すコマンド
-#
-#  # 追跡を外す       : git update-index --assume-unchanged docker/.env
-#  # 追跡を再開する   : git update-index --no-assume-unchanged docker/.env
-#
-#  ※ 上記はターミナルで実行してください。ここに書いても harmless です。
-#######################################################################
-
 # docker/.env
 # Mac
-#DATASET_PATH=/Volumes/OTO-SSD/SeniorThesis/DataSet
+DATASET_PATH=/Volumes/OTO-SSD/SeniorThesis/DataSet/Lung_dataset_original
 # Windows
-DATASET_PATH=D:/Lung_dataset_original
+# DATASET_PATH=D:/Lung_dataset_original
