+++ docker/docker-compose.gpu.yaml
@@ -19,7 +19,7 @@ services:
 
     volumes:
       - ../:/workspace
-      - ${DATASET_PATH}:/workspace/data/Lung_dataset_original
+      - ${DATASET_PATH}:/workspace/data
 
     working_dir: /workspace
     command: bash -c "jupyter notebook --allow-root --ip=0.0.0.0 --port=50000"
