+++ docker/docker-compose.yaml
@@ -17,8 +17,7 @@ services:
 
     volumes:
       - ../:/workspace                  # NOCTUA 全体を /workspace にマウント
-      #- ${DATASET_PATH}:/workspace/data # データセットを /workspace/data にマウント
-      - ${DATASET_PATH}/Lung_dataset_original:/workspace/data/Lung_dataset_original
+      - ${DATASET_PATH}:/workspace/data # データセットを /workspace/data にマウント
 
     working_dir: /workspace
     command: bash -c "jupyter notebook --allow-root --ip=0.0.0.0 --port=50000"
