+++ docker/requirements.txt
@@ -18,14 +18,3 @@ einops==0.6.0
 matplotlib==3.6.2
 numpy==1.23.4
 Pillow==9.3.0
-
-pandas
-japanize-matplotlib
-jupyter
-openpyxl
-opencv-python
-pillow
-ipython
-pylibjpeg
-
-tensorboard>=2.12.0
